Improve RAM efficiency in deduplication
This commit reflects a series of improvements on RAM efficiency while
trying to debug a problem. The problem *wasn't* about RAM, but I didn't
realize that yet. But the efficiency gains are substantial and allow
this pipeline to run on much smaller machines, so we keep them.

This commit should have been multiple smaller commits. Sorry!

Here are the notable changes:

1. Instead of storing an index->doc_id string array (with a string per
   paragraph, absolutely massive), we store a pair `book_ids` and
   `book_offsets`. The book_ids is just a list of book ids, and the
   offsets are the cumulative indices of paragraph starts. As all
   paragraphs are hashed, in order, in `hashes`, this allows one to
   determine the doc_id from the index in the hash array by a simple
   bisection search. This incurs a small computational overhead but
   saves ~80GB of RAM.

2. Delete the in memory hash array after it's been consumed. This can
   save a lot of RAM as well.

3. We update defaults. The max_bucket_size is 30,000 instead of 10,000
   now. This covers all of IB1 and seemed sensible. And we allocate 16GB
   for external sort instead of 4GB.

4. We add progress bars. (This code is entangled with other code, which
   is why I chose not to separate this commit into multiple commits
   after the fact).

5. To facilitate debugging, it is now possible to *not* delete
   intermediate temporary files and then to resume with these later.
   Ideally this will never need to be used, but I found it helpful while
   debugging an error. In particular, this keeps the externally sorted
   TSV file, which is the most important step.

6. For the CPP interface, use memory-view casting instead of relying on
   array.array to make zero-copy explicit.

7. Finally, in the CPP deduplication extension, we use int64 for array
   lengths instead of int32s. This *is* where the actual bug was. The
   reason is that we had 1.4B paragraphs, each having 2 uint64_t
   components forming a hash. The index into this hash array was a
   normal (32 bit) signed int and this overflowed when dealing with 2.8B
   elements. This caused silent segfault failures.

+++ extensions/simhash_cpp/simhash_bindings.cpp
@@ -20,8 +20,8 @@ namespace idi_simhash {
    int hamming_distance(uint64_t h1_lo, uint64_t h1_hi, uint64_t h2_lo, uint64_t h2_hi);
    bool are_near_duplicates(uint64_t h1_lo, uint64_t h1_hi, uint64_t h2_lo, uint64_t h2_hi, int threshold);
    std::vector<uint32_t> extract_bands(uint64_t hash_lo, uint64_t hash_hi);
-   std::vector<std::pair<int, int>> process_bucket(const std::vector<int>& doc_indices, const uint64_t* hashes, int threshold, int max_bucket_size);
-   std::vector<std::pair<int, int>> process_bucket_batch(const std::vector<std::vector<int>>& buckets, const uint64_t* hashes, int threshold, int max_bucket_size);
+   std::vector<std::pair<int64_t, int64_t>> process_bucket(const std::vector<int64_t>& doc_indices, const uint64_t* hashes, int threshold, int max_bucket_size);
+   std::vector<std::pair<int64_t, int64_t>> process_bucket_batch(const std::vector<std::vector<int64_t>>& buckets, const uint64_t* hashes, int threshold, int max_bucket_size);
 }
 
 /**
@@ -93,7 +93,7 @@ std::vector<uint32_t> py_extract_bands(py::object hash_value) {
  * The buffer should be an array of uint64_t values (2 per document: lo, hi).
  */
 py::list py_process_bucket(
-    const std::vector<int>& doc_indices,
+    const std::vector<int64_t>& doc_indices,
     py::buffer hashes_buffer,
     int threshold,
     int max_bucket_size
@@ -118,7 +118,7 @@ py::list py_process_bucket(
  * Process a batch of buckets using a buffer of hashes.
  */
 py::list py_process_bucket_batch(
-    const std::vector<std::vector<int>>& buckets,
+    const std::vector<std::vector<int64_t>>& buckets,
     py::buffer hashes_buffer,
     int threshold,
     int max_bucket_size
