+++ extensions/simhash_cpp/simhash_impl.cpp
@@ -519,30 +519,30 @@ namespace idi_simhash {
      * @param max_bucket_size  Skip buckets larger than this
      * @return Vector of (d1, d2) pairs that are duplicates
      */
-    std::vector<std::pair<int, int>> process_bucket(
-        const std::vector<int>& doc_indices,
+    std::vector<std::pair<int64_t, int64_t>> process_bucket(
+        const std::vector<int64_t>& doc_indices,
         const uint64_t* hashes,
         int threshold,
         int max_bucket_size
     ) {
-        std::vector<std::pair<int, int>> verified;
+        std::vector<std::pair<int64_t, int64_t>> verified;
 
         if (static_cast<int>(doc_indices.size()) > max_bucket_size) {
             return verified;  // Skip oversized buckets
         }
 
         // Sort indices for consistent ordering
-        std::vector<int> sorted_indices = doc_indices;
+        std::vector<int64_t> sorted_indices = doc_indices;
         std::sort(sorted_indices.begin(), sorted_indices.end());
 
         int n = static_cast<int>(sorted_indices.size());
         for (int i = 0; i < n; i++) {
-            int d1 = sorted_indices[i];
+            int64_t d1 = sorted_indices[i];
             uint64_t h1_lo = hashes[d1 * 2];
             uint64_t h1_hi = hashes[d1 * 2 + 1];
 
             for (int j = i + 1; j < n; j++) {
-                int d2 = sorted_indices[j];
+                int64_t d2 = sorted_indices[j];
                 uint64_t h2_lo = hashes[d2 * 2];
                 uint64_t h2_hi = hashes[d2 * 2 + 1];
 
@@ -564,13 +564,13 @@ namespace idi_simhash {
      * @param max_bucket_size  Skip buckets larger than this
      * @return Vector of (d1, d2) pairs that are duplicates
      */
-    std::vector<std::pair<int, int>> process_bucket_batch(
-        const std::vector<std::vector<int>>& buckets,
+    std::vector<std::pair<int64_t, int64_t>> process_bucket_batch(
+        const std::vector<std::vector<int64_t>>& buckets,
         const uint64_t* hashes,
         int threshold,
         int max_bucket_size
     ) {
-        std::vector<std::pair<int, int>> all_verified;
+        std::vector<std::pair<int64_t, int64_t>> all_verified;
 
         for (const auto& bucket : buckets) {
             auto verified = process_bucket(bucket, hashes, threshold, max_bucket_size);
